Make top-level Handler example runnable
The code in `handler_example_test.go` uses internal packages, so it
doesn't run on `pkg.go.dev`. (When pkgsite converts the example to a
`main` function, it's no longer in the `connect-go` module.)

This PR fixes the problem by switching from a whole-file example to a
one-function example. On balance, I think this is worse - it hides the
service _implementation_, which is more important than making the
example runnable.

@joshcarp, what do you think?

Fixes #433.
