+++ .golangci.yml
@@ -74,7 +74,7 @@ issues:
       path: protocol_grpc.go
     # We purposefully do an ineffectual assignment for an example.
     - linters: [ineffassign]
-      path: client_example_test.go
+      path: example_test.go
     # The generated file is effectively a global receiver.
     - linters: [varnamelen]
       path: cmd/protoc-gen-connect-go
@@ -84,10 +84,10 @@ issues:
       path: error_writer_example_test.go
     # It should be crystal clear that Connect uses plain *http.Clients.
     - linters: [revive, stylecheck]
-      path: client_example_test.go
+      path: example_test.go
     # Don't complain about timeout management or lack of output assertions in examples.
     - linters: [gosec, testableexamples]
-      path: handler_example_test.go
+      path: example_test.go
     # No output assertions needed for these examples.
     - linters: [testableexamples]
       path: error_writer_example_test.go
